+++ melanoma_app/package.json
@@ -7,7 +7,8 @@
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "ts:check": "tsc"
   },
   "dependencies": {
     "@expo/metro-config": "^0.10.7",
