+++ melanoma_app/src/components/home/help/tutorialsOverview.tsx
@@ -1,15 +1,17 @@
 import { Fragment } from "react";
 import { FlatList, ListRenderItemInfo, View, StyleSheet } from "react-native";
 
-import Styles from "../../styles";
-import { Tutorial, getTutorials } from "../../utils/helpData";
-import { InfoCard } from "../card";
+import { InfoCard } from "./cards";
+import Styles from "../../../styles";
+import { getTutorials } from "../../../utils/helpData";
 import {
   FollowupCardIcon,
   NotificationCardIcon,
   PhotoCardIcon,
   PrediagnosisCardIcon,
-} from "../icons";
+} from "../../icons";
+
+import ITutorial from "@/models/tutorial";
 
 const getIcon = (name: string) => {
   switch (name) {
@@ -38,19 +40,19 @@ const TutorialsOverview = (props: TutorialsOverviewProps) => {
     return label.includes(props.searchFilter);
   });
 
-  const renderCard = ({ item }: ListRenderItemInfo<Tutorial>) => {
+  const renderCard = ({ item }: ListRenderItemInfo<ITutorial>) => {
     return (
       <InfoCard
         title={item.title}
         backgroundColor={item.color}
         icon={getIcon(item.icon)}
-        href={`/help/tutorial/${item.id}`}
+        tutorialId={item.id}
       />
     );
   };
 
   const gap = () => {
-    return <View style={styles.gap} />;
+    return <View style={Styles.gap} />;
   };
 
   return (
@@ -69,9 +71,6 @@ const TutorialsOverview = (props: TutorialsOverviewProps) => {
 const styles = StyleSheet.create({
   center: {
     justifyContent: "center",
-  },
-  gap: {
-    width: 10,
   },
 });
