+++ melanoma_app/src/components/photo/cameraPreview.tsx
