+++ melanoma_app/src/models/question.ts
@@ -1,0 +1,5 @@
+export default interface IQuestion {
+  title: string;
+  body: string;
+  id: number;
+}
